New changes

+++ python/MuonSelector_cfi.py
@@ -1,11 +1,11 @@
 import FWCore.ParameterSet.Config as cms
 
-from Analysis.VLQAna.Muon_cfi import * 
+from MyAnalysis.VLQAna.Muon_cfi import * 
 
 defaultMuonSelectionParameters = cms.PSet(
     defaultMuonParameters, 
     muidtype = cms.string("TIGHT"), 
-    muPtMin = cms.double(0),
+    muPtMin = cms.double(20),
     muPtMax = cms.double(10000),
     muAbsEtaMax = cms.double(2.1),
     muCharge = cms.double(1.0),
