+++ python/ZCandSelector_cfi.py
@@ -1,8 +1,8 @@
 import FWCore.ParameterSet.Config as cms
 
 defaultZCandSelectionParameters = cms.PSet(
-    massMin = cms.double(60),
-    massMax = cms.double(120),
+    massMin = cms.double(20),
+    massMax = cms.double(100000),
     ptMin = cms.double(0),
     ptMax = cms.double(100000), 
     )
